👌 move things around
* split `helpers` to `browserHelpers` and `sdkHelpers`
* move `waitForServersIdle` to the `servers` module
* create a `testSetup` directory containing all code that is only used
for test setup
* move server applications into `testSetup/mockServerApp` and
`testSetup/intakeServerApp`
