+++ test/e2e/lib/browserHelpers.ts
@@ -1,23 +1,3 @@
-import { getTestServers } from './servers'
-
-export async function flushEvents() {
-  // wait to process user actions + event loop before switching page
-  await browserExecuteAsync((done) =>
-    setTimeout(() => {
-      done(undefined)
-    }, 200)
-  )
-  await waitForServersIdle()
-  const servers = await getTestServers()
-  await browser.url(`${servers.base.url}/empty`)
-  await waitForServersIdle()
-}
-
-export async function waitForServersIdle() {
-  const servers = await getTestServers()
-  return Promise.all([servers.base.waitForIdle(), servers.crossOrigin.waitForIdle(), servers.intake.waitForIdle()])
-}
-
 // typing issue for execute https://github.com/webdriverio/webdriverio/issues/3796
 export async function browserExecute(fn: any) {
   return browser.execute(fn as any)
